+++ libunhv3/unhv3.cpp
@@ -100,4 +100,3 @@ bool Unhv3::open(
     status = Unhv3Status::NO_ERROR;
     return true;
 }
-